+++ ankimorphs/__init__.py
@@ -18,6 +18,7 @@ from pathlib import Path
 from typing import Literal
 
 import aqt
+from anki import hooks
 from anki.cards import Card
 from anki.collection import OpChangesAfterUndo
 from aqt import gui_hooks, mw
@@ -51,8 +52,8 @@ from .ankimorphs_config import AnkiMorphsConfig, AnkiMorphsConfigFilter
 from .ankimorphs_db import AnkiMorphsDB
 from .extra_settings import ankimorphs_extra_settings, extra_settings_keys
 from .extra_settings.ankimorphs_extra_settings import AnkiMorphsExtraSettings
-from .filters import register_filters
 from .generators.generators_window import GeneratorWindow
+from .jit_highlight_morphs import am_highlight_morphs
 from .known_morphs_exporter import KnownMorphsExporterDialog
 from .progression.progression_window import ProgressionWindow
 from .recalc import recalc_main
@@ -97,6 +98,8 @@ def main() -> None:
     gui_hooks.state_did_undo.append(rebuild_seen_morphs)
 
     gui_hooks.profile_will_close.append(cleanup_profile_session)
+
+    hooks.field_filter.append(am_highlight_morphs)
 
 
 def init_toolbar_items(links: list[str], toolbar: Toolbar) -> None:
